https://guidewirejira.atlassian.net/browse/ISGOSU-508 - replace StringCache usage with StringPool - move StringPool to a different package and deprecate it - make StringPool.get(String) a no-op, simply return the String instead of caching it, this tests whether we can altogether remove StringPool strictly in terms of Gosu's use of it according to the effects on perf
