--- gosu-core-api/src/main/java/gw/internal/gosu/parser/StringCache.java
@@ -1,16 +1,0 @@
-/*
- * Copyright 2014 Guidewire Software, Inc.
- */
-
-package gw.internal.gosu.parser;
-
-import gw.util.StringPool;
-
-public class StringCache {
-  public static String get(String rawString) {
-    if (rawString == null) {
-      return null;
-    }
-    return StringPool.get( rawString );
-  }
-}
